caldav-pi.yaml: Radicale updates and improvements
- migrated to Radicale 2.x running on Py 3
- better nginx "routing" thanks to Radicale docs

+++ caldav-pi/start-radicale.sh
@@ -2,16 +2,16 @@
 
 PID_FILE={{ radicale_dir }}/radicale.pid
 
-pgrep -F $PID_FILE
-if [ $? -eq 0 ]; then
-    echo "Process already running."
-    exit 0
-else
-    rm -f $PID_FILE
+if [ -f $PID_FILE ]; then
+    pgrep -F $PID_FILE
+    if [ $? -eq 0 ]; then
+        echo "Process already running."
+        exit 0
+    else
+        rm -f $PID_FILE
+    fi
 fi
 
-source /home/{{ radicale_user }}/radicale-venv/bin/activate
-radicale \
+python3 -m radicale \
   --config {{ radicale_dir }}/radicale.conf \
   --pid $PID_FILE
-
